refactor(06): inject modules into index.js
